--- scoringrules/backend/arrayapi/__init__.py
@@ -1,4 +1,0 @@
-from .array_type import Array, Dtype
-from .functions import ArrayAPIProtocol
-
-__all__ = ["Array", "Dtype", "ArrayAPIProtocol"]
