--- scoringrules/new_backend/__init__.py
@@ -1,9 +1,0 @@
-from .registry import BackendsRegistry
-
-backends = BackendsRegistry()
-
-try:
-    import numba
-    _NUMBA_IMPORTED = True
-except:
-    _NUMBA_IMPORTED = False
